Allows passing in clients to useLazyQuery and useMutation Hooks
Update package version and add support for a new client generator in useLazyQuery and useMutation hooks.

This commit updates the package version to 0.2.3 and adds support for a new client generator in the useLazyQuery and useMutation hooks. The code changes include modifications to the `useLazyQuery.ts`, `useMutation.ts`, `useQuery.ts`, `index.ts` and `root.tsx` files, which allow users to pass a client generator function as an option when using these hooks. This enables greater flexibility in choosing which Apollo Client instance to use with these hooks.
Update package version and add support for a new client generator in useLazyQuery and useMutation hooks.

This commit updates the package version to 0.2.3 and adds support for a new client generator in the useLazyQuery and useMutation hooks. The code changes include modifications to the `useLazyQuery.ts`, `useMutation.ts`, `useQuery.ts`, `index.ts` and `root.tsx` files, which allow users to pass a client generator function as an option when using these hooks. This enables greater flexibility in choosing which Apollo Client instance to use with these hooks.